Only update alert phone number changes the logs when a change happens
Launching the settings screen always log an entry for the alert phone
number field. Ideally it should only log an entry when there has been
a change to the value of the field.

To solve this, compare the saved value in shared preferences with
the one currently entered in the field. If they don't match log
the entry