Provide request queue as singleton
It automatically detects which underling http client to use based on
the version of Android running on the device. Gingerbread and above
uses `HttpUrlConnection`. It also specifies the directory name to be
used by the disk based caching system of volley. Uses SMSsync version
name as part of the user agent label.