Fix #334 - Port Old Settings Implement to Use PreferenceFragmentCompat
Though it has eased some of the manual way of handling the state
of the settings controls it poses it own issues. Mostly they are
bugs because of the support library. For instance the fonts are huge
and had to implement a workaround to display a nice looking background
color.

Add a service based scope for DI to inject classes scoped for background service.
Remove all thirdparty libraries that aimed at providing support for preference
fragment.
Replaced all checkboxes with switch preference compat.