Fix broken search
Don't put navigation drawer's fragments to the back stack since we don't pop
them. Use the instance of the message fragment created when
navigating to the message fragment. This way we use the same instance
to for performing a search query.

Make the search widget visible when user navigates to the incoming screen. It
was set to gone making it not show