Remove ability to track sync state
With the current UI and work flow, we no longer
track synchronization of pending messages' state

Remove `SyncPendingMesageState` class as it's no longer needed.

Remove the ability to check if a service is running as it was being
passed to the state class which is no longer relevant. Addresses
the issue raised in #411